+++ src/error-messages.ts
@@ -1,6 +1,4 @@
 
-export const mutatorWrongReturnType = "mutator methods must return void|undefined";
-
 export const illegalAccessToThis = `
 Illegal access to 'this'. Make sure that 'this' is not accessed in positions like the following: