+++ src/mutators.test.ts
@@ -1,6 +1,8 @@
 import {enableTyduxDevelopmentMode} from "./development";
 import {Mutators} from "./mutators";
 import {createSimpleStore} from "./SimpleStore";
+import {Store} from "./Store";
+import {collect, createAsyncPromise} from "./test-utils";
 
 
 describe("Mutators", function () {
@@ -17,7 +19,7 @@ describe("Mutators", function () {
         }
 
         const store = createSimpleStore("", new TestMutator(), {n1: 0});
-        store.mutate.mod1();
+        store.mod1();
         assert.deepEqual(store.state, {n1: 1});
     });
 
@@ -31,7 +33,7 @@ describe("Mutators", function () {
         }
 
         const store = createSimpleStore("", new TestMutator(), {n1: 0});
-        store.mutate.mod1();
+        store.mod1();
         assert.deepEqual(store.state, {n1: 1});
     });
 
@@ -53,7 +55,7 @@ describe("Mutators", function () {
         }
 
         const store = createSimpleStore("", new TestMutator(), {n1: ""});
-        store.mutate.mod1();
+        store.mod1();
         assert.deepEqual(store.state, {n1: "123"});
     });
 
@@ -69,8 +71,38 @@ describe("Mutators", function () {
         }
 
         const store = createSimpleStore("", new TestMutator(), {a: 0});
-        assert.throws(() => store.mutate.mut1());
+        assert.throws(() => store.mut1());
         assert.equal(store.state.a, 0);
+    });
+
+    it("member method can use member variables", function (done) {
+        class MyMutators extends Mutators<any> {
+
+            counterA?: number;
+
+            counterB = 1;
+
+            method() {
+                this.counterA = 10;
+                this.innerMethod();
+            }
+
+            innerMethod() {
+                this.counterB = 20;
+            }
+        }
+
+        class MyStore extends Store<MyMutators, any> {
+            action() {
+                this.mutate.method();
+                assert.equal(this.mutate.counterA, 10);
+                assert.equal(this.mutate.counterB, 20);
+                done();
+            }
+        }
+
+        const store = new MyStore("myStore", new MyMutators(), {});
+        store.action();
     });
 
 });